+++ internal/driver/gomobile/driver.go
@@ -26,6 +26,7 @@ const tapSecondaryDelay = 300 * time.Millisecond
 
 type mobileDriver struct {
 	app   app.App
+	quit  bool
 	glctx gl.Context
 
 	windows []fyne.Window
@@ -70,7 +71,7 @@ func (d *mobileDriver) CanvasForObject(fyne.CanvasObject) fyne.Canvas {
 		return nil
 	}
 
-	// TODO figure out how we handle multiple windows...
+	// TODO don't just assume it refers to the topmost window
 	return d.currentWindow().Canvas()
 }
 
@@ -94,14 +95,13 @@ func (d *mobileDriver) Quit() {
 		return
 	}
 
-	// TODO? often mobile apps should not allow this...
-	d.app.Send(lifecycle.Event{From: lifecycle.StageAlive, To: lifecycle.StageDead, DrawContext: nil})
+	// TODO should this be disabled for iOS?
+	d.quit = true
 }
 
 func (d *mobileDriver) Run() {
 	app.Main(func(a app.App) {
 		d.app = a
-		quit := false
 
 		var currentSize size.Event
 		for e := range a.Events() {
@@ -123,7 +123,7 @@ func (d *mobileDriver) Run() {
 					d.glctx = nil
 				}
 				if e.Crosses(lifecycle.StageAlive) == lifecycle.CrossOff {
-					quit = true
+					d.quit = true
 				}
 			case size.Event:
 				currentSize = e
@@ -157,7 +157,7 @@ func (d *mobileDriver) Run() {
 				}
 			}
 
-			if quit {
+			if d.quit {
 				break
 			}
 		}